Squashed 'src/cppgoslin/' changes from 8968a7897..0deae6c04
0deae6c04 fixing warnings

git-subtree-dir: src/cppgoslin
git-subtree-split: 0deae6c048aa5b878ff188a86afa6b7f56d23273

+++ src/parser/LipidBaseParserEventHandler.cpp
@@ -69,7 +69,7 @@ bool LipidBaseParserEventHandler::check_full_structure(FunctionalGroup *obj){
     if (is_fa && !obj->double_bonds->double_bond_positions.empty()){
         int sum = 0;
         for (auto &kv : obj->double_bonds->double_bond_positions) sum += kv.second == "E" || kv.second == "Z" || (kv.second == "" && kv.first == ((FattyAcid*)obj)->num_carbon - 1);
-        full &= sum == obj->double_bonds->double_bond_positions.size();
+        full &= sum == (int)obj->double_bonds->double_bond_positions.size();
         
     }
